+++ src/js/mixin/slider-autoplay.js
@@ -54,7 +54,9 @@ export default {
                 if (
                     !(
                         this.stack.length ||
-                        (this.draggable && matches(this.$el, ':focus-within')) ||
+                        (this.draggable &&
+                            matches(this.$el, ':focus-within') &&
+                            !matches(this.$el, ':focus')) ||
                         (this.pauseOnHover && matches(this.$el, ':hover'))
                     )
                 ) {